+++ src/PaperServiceProvider.php
@@ -16,6 +16,7 @@ use JacobJoergensen\LaravelPaper\Console\WarmCommand;
 use JacobJoergensen\LaravelPaper\Drivers\DriverRegistry;
 use JacobJoergensen\LaravelPaper\Drivers\JsonDriver;
 use JacobJoergensen\LaravelPaper\Drivers\MarkdownDriver;
+use JacobJoergensen\LaravelPaper\Drivers\YamlDriver;
 
 final class PaperServiceProvider extends ServiceProvider
 {
@@ -41,11 +42,13 @@ final class PaperServiceProvider extends ServiceProvider
 
         $this->app->singleton(MarkdownDriver::class);
         $this->app->singleton(JsonDriver::class);
+        $this->app->singleton(YamlDriver::class);
 
         $this->app->singleton(DriverRegistry::class, function (): DriverRegistry {
             $registry = new DriverRegistry;
             $registry->register('markdown', MarkdownDriver::class);
             $registry->register('json', JsonDriver::class);
+            $registry->register('yaml', YamlDriver::class);
 
             return $registry;
         });